requirements.txt
flask
gunicorn
opencv-python
numpy
Pillow
